requirements.txt
async-timeout==4.0.3
blinker==1.6.2
click==8.1.7
Flask==2.3.2
ijson==3.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
redis==5.0.0
shapely==2.0.1
typing==3.7.4.3
Werkzeug==2.3.7
